% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_data_extract.R
\name{ff_data_extract}
\alias{ff_data_extract}
\title{Extract data from downloaded file}
\usage{
ff_data_extract(file)
}
\arguments{
\item{file}{character string, file connection}
}
\description{
Internal function
}
