% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_list_available_files.R
\name{ff_list_available_files}
\alias{ff_list_available_files}
\title{List Fama-French files}
\usage{
ff_list_available_files(file_extension = c("TXT", "CSV"))
}
\arguments{
\item{file_extension}{character, FF supply CSV or TXT data.}
}
\value{
character vector of filenames available from the website
}
\description{
List Fama-French files
}
