% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_data_download.R
\name{ff_tidy_data}
\alias{ff_tidy_data}
\title{Fama-French data download}
\usage{
ff_tidy_data(file_name, all_subsets = FALSE)
}
\arguments{
\item{file_name}{name of FF zip file}

\item{all_subsets}{logical, if TRUE, returns all data subdivisions from file, if FALSE, only the first table (i.e. until first line-break after data)}
}
\value{
*tidy* \code{data.frame} with return data. Columns date_start, date_end are of type \code{Date}, columns series_name, series_type, data_type are \code{character}; column value is \code{numeric}.
}
\description{
Downloads data from \url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}, cleans, returns nice form.
}
